# .schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "runtime",
    "description": "The spiced runtime",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.2.0-unstable"
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local development server. Configure with `--http`."
    }
  ],
  "paths": {
...
    "/v1/tools/{name}": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Run Tool",
        "description": "The request body and JSON response formats match the tool’s specification.",
        "operationId": "post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the tool",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Tool specific input parameters. See /v1/tools for parameter schema.",
          "content": {
            "application/json": {
              "schema": {},
              "example": {
                "query": "SELECT avg(total_amount), avg(tip_amount), count(1), passenger_count FROM my_table GROUP BY passenger_count ORDER BY passenger_count ASC LIMIT 3"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tool Specific response, in JSON format",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "sql": {
                    "value": [
                      {
                        "AVG(my_table.tip_amount)": 3.072259971396793,
                        "AVG(my_table.total_amount)": 25.327816939456525,
                        "COUNT(Int64(1))": 31465,
                        "passenger_count": 0
                      },
                      {
                        "AVG(my_table.tip_amount)": 3.3712622884680057,
                        "AVG(my_table.total_amount)": 26.205230445474996,
                        "COUNT(Int64(1))": 2188739,
                        "passenger_count": 1
                      },
                      {
                        "AVG(my_table.tip_amount)": 3.7171302113290854,
                        "AVG(my_table.total_amount)": 29.520659930930304,
                        "COUNT(Int64(1))": 405103,
                        "passenger_count": 2
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tool not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "Tool no_sql not found"
              }
            }
          },
          "500": {
            "description": "Error occured whilst calling the tool",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "message": "Error calling tool no_sql: No such tool"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "security": [
    {
      "api_key": []
    }
  ]
}